+++ packages/react-dev-overlay/src/internal/components/CodeFrame/CodeFrame.tsx
@@ -2,7 +2,7 @@ import Anser from 'anser'
 import * as React from 'react'
 import { StackFrame } from 'stacktrace-parser'
 import stripAnsi from 'strip-ansi'
-import { getFrameSource } from '../../overlay/RuntimeErrors/utils'
+import { getFrameSource } from '../../helpers/stack-frame'
 
 export type CodeFrameProps = { stackFrame: StackFrame; codeFrame: string }
