+++ packages/react-dev-overlay/src/internal/helpers/stack-frame.ts
